docmind/requirements.txt
numpy==1.26.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sentence-transformers==3.0.1
torch
PyMuPDF==1.24.5
langchain-text-splitters==0.2.2
openai==1.35.7
streamlit==1.36.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0